+++ letsencrypt/client/tests/challenge_util_test.py
@@ -6,10 +6,11 @@ import unittest
 
 import M2Crypto
 
+from letsencrypt.acme import jose
+
 from letsencrypt.client import challenge_util
 from letsencrypt.client import client
 from letsencrypt.client import CONFIG
-from letsencrypt.client import le_util
 
 
 class DvsniGenCertTest(unittest.TestCase):
@@ -20,7 +21,7 @@ class DvsniGenCertTest(unittest.TestCase):
         """Basic test for straightline code."""
         domain = "example.com"
         dvsni_r = "r_value"
-        r_b64 = le_util.jose_b64encode(dvsni_r)
+        r_b64 = jose.b64encode(dvsni_r)
         pem = pkg_resources.resource_string(
             __name__, os.path.join("testdata", "rsa256_key.pem"))
         key = client.Client.Key("path", pem)
@@ -29,7 +30,7 @@ class DvsniGenCertTest(unittest.TestCase):
 
         # pylint: disable=protected-access
         ext = challenge_util._dvsni_gen_ext(
-            dvsni_r, le_util.jose_b64decode(s_b64))
+            dvsni_r, jose.b64decode(s_b64))
         self._standard_check_cert(cert_pem, domain, nonce, ext)
 
     def _standard_check_cert(self, pem, domain, nonce, ext):
